scrigoPackageVariables now holds int16 instead of int8

+++ internal/compiler/emitter_util.go
@@ -161,7 +161,7 @@ func mayHaveDepencencies(variables, values []ast.Expression) bool {
 	return !allDifferentIdentifiers()
 }
 
-func (e *emitter) predefinedVariableIndex(varRv reflect.Value) int8 {
+func (e *emitter) predefinedVariableIndex(varRv reflect.Value) int16 {
 	/*
 		if reflect.TypeOf(value).Kind() == reflect.Ptr {
 			e.globals = append(e.globals, vm.Global{Pkg: parserPredefinedPkg.Name, Name: ident, Value: value})
@@ -176,10 +176,10 @@ func (e *emitter) predefinedVariableIndex(varRv reflect.Value) int8 {
 	if ok {
 		return index
 	}
-	index = int8(len(e.globals))
+	index = int16(len(e.globals))
 	g := vm.Global{Pkg: "???", Name: "???", Value: varRv.Interface()}
 	if e.predefVarIndex[e.fb.fn] == nil {
-		e.predefVarIndex[e.fb.fn] = make(map[reflect.Value]int8)
+		e.predefVarIndex[e.fb.fn] = make(map[reflect.Value]int16)
 	}
 	e.globals = append(e.globals, g)
 	e.predefVarIndex[e.fb.fn][varRv] = index
